chore(firebase.ts): パスを変更
